+++ databento/__init__.py
@@ -1,6 +1,8 @@
 import logging
 import warnings
 
+from databento_dbn import BBOMsg
+from databento_dbn import CBBOMsg
 from databento_dbn import Compression
 from databento_dbn import Encoding
 from databento_dbn import ErrorMsg
@@ -14,11 +16,13 @@ from databento_dbn import OHLCVMsg
 from databento_dbn import Schema
 from databento_dbn import StatMsg
 from databento_dbn import StatType
+from databento_dbn import StatusMsg
 from databento_dbn import SType
 from databento_dbn import SymbolMappingMsg
 from databento_dbn import SystemMsg
 from databento_dbn import TradeMsg
 
+from databento.common import API_VERSION
 from databento.common import bentologging
 from databento.common import symbology
 from databento.common.dbnstore import DBNStore
@@ -26,6 +30,7 @@ from databento.common.enums import Delivery
 from databento.common.enums import FeedMode
 from databento.common.enums import HistoricalGateway
 from databento.common.enums import Packaging
+from databento.common.enums import ReconnectPolicy
 from databento.common.enums import RecordFlags
 from databento.common.enums import RollRule
 from databento.common.enums import SplitDuration
@@ -39,9 +44,9 @@ from databento.common.publishers import Publisher
 from databento.common.publishers import Venue
 from databento.common.symbology import InstrumentMap
 from databento.common.types import DBNRecord
-from databento.historical.api import API_VERSION
 from databento.historical.client import Historical
 from databento.live.client import Live
+from databento.reference.client import Reference
 from databento.version import __version__  # noqa
 
 
@@ -69,26 +74,31 @@ __all__ = [
     "HistoricalGateway",
     "InstrumentMap",
     "Live",
+    "Reference",
     "Packaging",
+    "ReconnectPolicy",
     "RollRule",
     "Schema",
     "SplitDuration",
+    "StatType",
     "SType",
     "SymbologyResolution",
     # DBN Record Types
     "Metadata",
+    "BBOMsg",
+    "CBBOMsg",
+    "ErrorMsg",
+    "ImbalanceMsg",
+    "InstrumentDefMsg",
     "MBOMsg",
     "MBP1Msg",
     "MBP10Msg",
-    "TradeMsg",
     "OHLCVMsg",
     "StatMsg",
-    "InstrumentDefMsg",
-    "ImbalanceMsg",
-    "ErrorMsg",
-    "StatType",
-    "SystemMsg",
+    "StatusMsg",
     "SymbolMappingMsg",
+    "SystemMsg",
+    "TradeMsg",
 ]
 
 # Setup logging