+++ databento/historical/api/symbology.py
@@ -7,14 +7,14 @@ from typing import Any
 from databento_dbn import SType
 from requests import Response
 
+from databento.common import API_VERSION
+from databento.common.http import BentoHttpAPI
 from databento.common.parsing import datetime_to_date_string
 from databento.common.parsing import optional_date_to_string
 from databento.common.parsing import optional_symbols_list_to_list
 from databento.common.publishers import Dataset
 from databento.common.validation import validate_enum
 from databento.common.validation import validate_semantic_string
-from databento.historical.api import API_VERSION
-from databento.historical.http import BentoHttpAPI
 
 
 class SymbologyHttpAPI(BentoHttpAPI):